+++ src/App/NetDaemon.App/Common/Fluent/Fluent.cs
@@ -60,42 +60,6 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
     { }
 
     /// <summary>
-    ///     Properties on entities that can be filtered in lambda expression
-    /// </summary>
-    public interface IEntityProperties
-    {
-        /// <summary>
-        ///     Filter on area where the entity device are at
-        /// </summary>
-        string? Area { get; set; }
-
-        /// <summary>
-        ///     Filter on attribute
-        /// </summary>
-        dynamic? Attribute { get; set; }
-
-        /// <summary>
-        ///     Filter on unique id of the entity
-        /// </summary>
-        string EntityId { get; set; }
-
-        /// <summary>
-        ///     Filter on last changed time
-        /// </summary>
-        DateTime LastChanged { get; set; }
-
-        /// <summary>
-        ///     Filter on last updated time
-        /// </summary>
-        DateTime LastUpdated { get; set; }
-
-        /// <summary>
-        ///     Filter on state
-        /// </summary>
-        dynamic? State { get; set; }
-    }
-
-    /// <summary>
     ///     Expressions that ends with execute
     /// </summary>
     public interface IExecute
@@ -107,7 +71,7 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
     }
 
     /// <summary>
-    ///     Expressions that ends with async excecution
+    ///     Expressions that ends with async execution
     /// </summary>
     public interface IExecuteAsync
     {
@@ -123,7 +87,7 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
     public interface IScript
     {
         /// <summary>
-        ///     Excecutes scripts async
+        ///     Executes scripts async
         /// </summary>
         Task ExecuteAsync();
     }
